LIKA-127: Add users allowed to create other users to config
