+++ ckanext/ckanext-apicatalog_scheming/ckanext/apicatalog_scheming/plugin.py
@@ -6,18 +6,12 @@ from ckan.common import config
 import ckan.lib.navl.dictization_functions as dictization_functions
 
 import validators
-from schema import create_user_to_organization_schema
-
-ValidationError = toolkit.ValidationError
-get_action = toolkit.get_action
-_validate = dictization_functions.validate
 
 
 class Apicatalog_SchemingPlugin(plugins.SingletonPlugin):
     plugins.implements(plugins.IConfigurer)
     plugins.implements(plugins.IValidators)
     plugins.implements(plugins.ITemplateHelpers)
-    plugins.implements(plugins.IActions)
 
     # IConfigurer
 
@@ -39,39 +33,6 @@ class Apicatalog_SchemingPlugin(plugins.SingletonPlugin):
 
     def get_helpers(self):
         return {'scheming_field_only_default_required': scheming_field_only_default_required}
-
-    def get_actions(self):
-        return {
-            "create_user_to_organization": create_user_to_organization
-        }
-
-def create_user_to_organization(context, data_dict):
-
-    def _generate_password():
-        out = ''
-        for n in xrange(8):
-            out += str(uuid.uuid4())
-        return out
-
-    model = context['model']
-    schema = context.get('schema') or create_user_to_organization_schema()
-    session = context['session']
-
-    data, errors = _validate(data_dict, schema, context)
-
-    if errors:
-        session.rollback()
-        raise ValidationError(errors)
-
-    data['password'] = _generate_password()
-    user = get_action('user_create')(context, data)
-    return {
-        "name": user['name'] ,
-        "email": user['email']
-    }
-
-
-
 
 
 def scheming_field_only_default_required(field, lang):
