+++ ckanext/ckanext-apicatalog_scheming/ckanext/apicatalog_scheming/schema.py
@@ -6,6 +6,7 @@ def create_user_to_organization_schema(not_empty, name_validator, user_name_vali
     return {
         "name": [not_empty, name_validator, user_name_validator, unicode_safe],
         "email": [not_empty, unicode_safe, email_validator],
-        "business_id": [not_empty, unicode_safe, business_id_validator]
+        "business_id": [not_empty, unicode_safe, business_id_validator],
+        "organization_name": [not_empty, unicode_safe]
     }
